Calculo del saldo del cliente terminado

+++ ms-pedidos/src/main/java/isi/dan/ms/pedidos/modelo/DetallePedido.java
@@ -12,7 +12,7 @@ public class DetallePedido {
     private Producto producto;
     private Integer cantidad;
     private BigDecimal precioUnitario;
-    private BigDecimal descuento;
+    private Integer descuento;
     private BigDecimal precioFinal;
